Added the option to control GR deviation with data passed to model.  Slight deviation from GR is preferred.
Also added time-offset H-L that is marginalized over (holding the antenna
patters, inclination, etc, fixed, since we are just modifying the "narrow"
dimension of the sky pattern).

Still finding some pesky residual power in L that *just won't go away*.
Probably a bug, but for the life of me, cannot find it.